requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-jose
passlib
email-validator